--- src/main/java/org/openjdk/jextract/impl/Constants.java
@@ -1,530 +1,0 @@
-/*
- * Copyright (c) 2021, 2023, Oracle and/or its affiliates. All rights reserved.
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- *
- * This code is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation.  Oracle designates this
- * particular file as subject to the "Classpath" exception as provided
- * by Oracle in the LICENSE file that accompanied this code.
- *
- * This code is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
- * version 2 for more details (a copy is included in the LICENSE file that
- * accompanied this code).
- *
- * You should have received a copy of the GNU General Public License version
- * 2 along with this work; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
- * or visit www.oracle.com if you need additional information or have any
- * questions.
- */
-
-package org.openjdk.jextract.impl;
-
-import org.openjdk.jextract.Declaration;
-import org.openjdk.jextract.Type;
-
-import javax.tools.JavaFileObject;
-import java.lang.foreign.AddressLayout;
-import java.lang.foreign.FunctionDescriptor;
-import java.lang.foreign.GroupLayout;
-import java.lang.foreign.MemoryLayout;
-import java.lang.foreign.MemorySegment;
-import java.lang.foreign.SequenceLayout;
-import java.lang.foreign.StructLayout;
-import java.lang.foreign.ValueLayout;
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.VarHandle;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.StringJoiner;
-import java.util.function.Function;
-
-public class Constants {
-
-    private final Map<Object, Constant> cache = new HashMap<>();
-
-    private final List<SourceFileBuilder> constantBuilders = new ArrayList<>();
-    private Builder currentBuilder;
-
-    private final String packageName;
-
-    public Constants(String packageName) {
-        this.packageName = packageName;
-        currentBuilder = startNewBuilder(packageName, 0);
-        // prime the cache with basic primitive/pointer (immediate) layouts
-        for (Type.Primitive.Kind kind : Type.Primitive.Kind.values()) {
-            kind.layout().ifPresent(layout -> {
-                if (layout instanceof ValueLayout valueLayout) {
-                    cache.put(valueLayout, ImmediateConstant.ofPrimitiveLayout(valueLayout));
-                }
-            });
-        }
-        AddressLayout pointerLayout = ValueLayout.ADDRESS.withTargetLayout(
-                MemoryLayout.sequenceLayout(Long.MAX_VALUE, ValueLayout.JAVA_BYTE));
-        cache.put(pointerLayout, ImmediateConstant.ofPrimitiveLayout(pointerLayout));
-    }
-
-    static final int CONSTANTS_PER_CLASS = Integer.getInteger("jextract.constants.per.class", 5);
-
-    private Builder startNewBuilder(String packageName, int id) {
-        String builderClassName = "constants$" + id;
-        SourceFileBuilder sfb = SourceFileBuilder.newSourceFile(packageName, builderClassName);
-        constantBuilders.add(sfb);
-        Builder builder = new Builder(sfb, sfb.className());
-        builder.classBegin();
-        builder.emitPrivateDefaultConstructor();
-        return builder;
-    }
-
-    private Builder builder() {
-        if (currentBuilder.constantIndex > CONSTANTS_PER_CLASS) {
-            currentBuilder.classEnd();
-            currentBuilder = startNewBuilder(packageName, constantBuilders.size());
-        }
-        return currentBuilder;
-    }
-
-    static sealed abstract class Constant permits Builder.NamedConstant, ImmediateConstant {
-
-        final Class<?> type;
-
-        public Constant(Class<?> type) {
-            this.type = type;
-        }
-
-        Class<?> type() {
-            return type;
-        }
-
-        String getterName(String javaName) {
-            return javaName + nameSuffix();
-        }
-
-        Constant emitGetter(ClassSourceBuilder builder, String mods, String javaName) {
-            return emitGetter(builder, mods, c -> c.getterName(javaName));
-        }
-
-        Constant emitGetter(ClassSourceBuilder builder, String mods, String javaName, String symbolName) {
-            return emitGetter(builder, mods, symbolName, c -> c.getterName(javaName));
-        }
-
-        Constant emitGetterWithComment(ClassSourceBuilder builder, String mods, String javaName, String symbolName,
-                                       Declaration decl) {
-            return emitGetterWithComment(builder, mods, symbolName, c -> c.getterName(javaName), decl);
-        }
-
-        Constant emitGetter(ClassSourceBuilder builder, String mods, Function<Constant, String> getterNameFunc) {
-            builder.emitConstantGetter(mods, getterNameFunc.apply(this), false, null, this, null);
-            return this;
-        }
-
-        Constant emitGetterWithComment(ClassSourceBuilder builder, String mods, Function<Constant, String> getterNameFunc,
-                                       Declaration decl) {
-            builder.emitConstantGetter(mods, getterNameFunc.apply(this), false, null, this, decl);
-            return this;
-        }
-
-        Constant emitGetter(ClassSourceBuilder builder, String mods, String symbolName, Function<Constant, String> getterNameFunc) {
-            builder.emitConstantGetter(mods, getterNameFunc.apply(this), true, symbolName, this, null);
-            return this;
-        }
-
-        Constant emitGetterWithComment(ClassSourceBuilder builder, String mods, String symbolName,
-                                       Function<Constant, String> getterNameFunc, Declaration decl) {
-            builder.emitConstantGetter(mods, getterNameFunc.apply(this), true, symbolName, this, decl);
-            return this;
-        }
-
-        String nameSuffix() {
-            if (type.equals(MemorySegment.class)) {
-                return "$SEGMENT";
-            } else if (type.equals(MemoryLayout.class)) {
-                return "$LAYOUT";
-            } else if (type.equals(MethodHandle.class)) {
-                return "$MH";
-            } else if (type.equals(VarHandle.class)) {
-                return "$VH";
-            } else if (type.equals(FunctionDescriptor.class)) {
-                return "$DESC";
-            } else {
-                return "";
-            }
-        }
-
-        abstract String accessExpression();
-
-        @Override
-        public String toString() {
-            return accessExpression();
-        }
-    }
-
-    final static class ImmediateConstant extends Constant {
-        final String value;
-
-        ImmediateConstant(Class<?> type, String value) {
-            super(type);
-            this.value = value;
-        }
-
-        @Override
-        String accessExpression() {
-            return value;
-        }
-
-        static ImmediateConstant ofPrimitiveLayout(ValueLayout vl) {
-            final String layoutStr;
-            if (vl.carrier() == boolean.class) {
-                layoutStr = "JAVA_BOOLEAN";
-            } else if (vl.carrier() == char.class) {
-                layoutStr = "JAVA_CHAR";
-            } else if (vl.carrier() == byte.class) {
-                layoutStr = "JAVA_BYTE";
-            } else if (vl.carrier() == short.class) {
-                layoutStr = "JAVA_SHORT";
-            } else if (vl.carrier() == int.class) {
-                layoutStr = "JAVA_INT";
-            } else if (vl.carrier() == float.class) {
-                layoutStr = "JAVA_FLOAT";
-            } else if (vl.carrier() == long.class) {
-                layoutStr = "JAVA_LONG";
-            } else if (vl.carrier() == double.class) {
-                layoutStr = "JAVA_DOUBLE";
-            } else if (vl.carrier() == MemorySegment.class) {
-                layoutStr = "RuntimeHelper.POINTER";
-            } else {
-                throw new UnsupportedOperationException("Unsupported layout: " + vl);
-            }
-            return new ImmediateConstant(MemoryLayout.class, layoutStr);
-        }
-
-        static Constant ofLiteral(Class<?> type, Object value) {
-            StringBuilder buf = new StringBuilder();
-            if (type == float.class) {
-                float f = ((Number)value).floatValue();
-                if (Float.isFinite(f)) {
-                    buf.append(value);
-                    buf.append("f");
-                } else {
-                    buf.append("Float.valueOf(\"");
-                    buf.append(value);
-                    buf.append("\")");
-                }
-            } else if (type == long.class) {
-                buf.append(value.toString());
-                buf.append("L");
-            } else if (type == double.class) {
-                double d = ((Number)value).doubleValue();
-                if (Double.isFinite(d)) {
-                    buf.append(value);
-                    buf.append("d");
-                } else {
-                    buf.append("Double.valueOf(\"");
-                    buf.append(value);
-                    buf.append("\")");
-                }
-            } else if (type == boolean.class) {
-                boolean booleanValue = ((Number)value).byteValue() != 0;
-                buf.append(booleanValue);
-            } else {
-                buf.append("(" + type.getName() + ")");
-                buf.append(value + "L");
-            }
-            return new ImmediateConstant(type, buf.toString());
-        }
-    }
-
-    public List<JavaFileObject> toFiles() {
-        currentBuilder.classEnd();
-        return new ArrayList<>(constantBuilders.stream()
-                .map(SourceFileBuilder::toFile).toList());
-    }
-
-    class Builder extends ClassSourceBuilder {
-
-        private static final String MEMBER_MODS = "static final";
-
-        Builder(SourceFileBuilder builder, String className) {
-            super(builder, "final", Kind.CLASS, className, null, null);
-        }
-
-        int constantIndex = 0;
-
-        final class NamedConstant extends Constant {
-            final String constantName;
-
-            NamedConstant(Class<?> type) {
-                super(type);
-                this.constantName = newConstantName();
-            }
-
-            @Override
-            String accessExpression() {
-                return className() + "." + constantName;
-            }
-        }
-
-        private Constant emitDowncallMethodHandleField(String nativeName, FunctionDescriptor descriptor, boolean isVarargs, boolean virtual) {
-            Constant functionDesc = addFunctionDesc(descriptor);
-            incrAlign();
-            NamedConstant mhConst = new NamedConstant(MethodHandle.class);
-            indent();
-            append(MEMBER_MODS + " MethodHandle ");
-            append(mhConst.constantName + " = RuntimeHelper.");
-            if (isVarargs) {
-                append("downcallHandleVariadic");
-            } else {
-                append("downcallHandle");
-            }
-            append("(\n");
-            incrAlign();
-            indent();
-            if (!virtual) {
-                append("\"" + nativeName + "\"");
-                append(",\n");
-                indent();
-            }
-            append(functionDesc.accessExpression());
-            append("\n");
-            decrAlign();
-            indent();
-            append(");\n");
-            decrAlign();
-            return mhConst;
-        }
-
-        private Constant emitUpcallMethodHandleField(String className, String methodName, FunctionDescriptor descriptor) {
-            Constant functionDesc = addFunctionDesc(descriptor);
-            NamedConstant mhConst = new NamedConstant(MethodHandle.class);
-            appendIndentedLines(STR."""
-                static final MethodHandle \{mhConst.constantName} = RuntimeHelper.upcallHandle(\{className}.class, "\{methodName}", \{functionDesc});
-                """);
-            return mhConst;
-        }
-
-        private Constant emitVarHandle(ValueLayout valueLayout) {
-            Constant layoutConstant = addLayout(valueLayout);
-            NamedConstant vhConst = new NamedConstant(VarHandle.class);
-            appendIndentedLines(STR."""
-                static final VarHandle \{vhConst.constantName} = \{layoutConstant}.varHandle();
-                """);
-            return vhConst;
-        }
-
-        private static String pathElementStr(String nativeName, List<String> prefixElementNames) {
-            StringJoiner joiner = new StringJoiner(", ");
-            for (String prefixElementName : prefixElementNames) {
-                joiner.add(STR."MemoryLayout.PathElement.groupElement(\"\{prefixElementName}\")");
-            }
-            joiner.add(STR."MemoryLayout.PathElement.groupElement(\"\{nativeName}\")");
-            return joiner.toString();
-        }
-
-        private Constant emitFieldVarHandle(String nativeName, GroupLayout parentLayout, List<String> prefixElementNames) {
-            Constant layoutConstant = addLayout(parentLayout);
-            NamedConstant vhConst = new NamedConstant(VarHandle.class);
-            appendIndentedLines(STR."""
-                static final VarHandle \{vhConst.constantName} = \{layoutConstant}.varHandle(\{pathElementStr(nativeName, prefixElementNames)});
-                """);
-            return vhConst;
-        }
-
-        private Constant emitLayoutField(MemoryLayout layout) {
-            NamedConstant layoutConst = new NamedConstant(MemoryLayout.class);
-            incrAlign();
-            indent();
-            String layoutClassName = Utils.layoutDeclarationType(layout).getSimpleName();
-            append(STR."static final \{layoutClassName} \{layoutConst.constantName} = ");
-            emitLayoutString(layout);
-            append(";\n");
-            decrAlign();
-            return layoutConst;
-        }
-
-        private void emitLayoutString(MemoryLayout l) {
-            if (l instanceof ValueLayout val) {
-                append(ImmediateConstant.ofPrimitiveLayout(val).accessExpression());
-                if (l.byteAlignment() != l.byteSize()) {
-                    append(STR.".withByteAlignment(\{l.byteAlignment()})");
-                }
-            } else if (l instanceof SequenceLayout seq) {
-                append(STR."MemoryLayout.sequenceLayout(\{seq.elementCount()}, ");
-                emitLayoutString(seq.elementLayout());
-                append(")");
-            } else if (l instanceof GroupLayout group) {
-                if (group instanceof StructLayout) {
-                    append("MemoryLayout.structLayout(\n");
-                } else {
-                    append("MemoryLayout.unionLayout(\n");
-                }
-                incrAlign();
-                String delim = "";
-                for (MemoryLayout e : group.memberLayouts()) {
-                    append(delim);
-                    indent();
-                    emitLayoutString(e);
-                    delim = ",\n";
-                }
-                append("\n");
-                decrAlign();
-                indent();
-                append(")");
-            } else {
-                // padding (or unsupported)
-                append(STR."MemoryLayout.paddingLayout(\{l.byteSize()})");
-            }
-            if (l.name().isPresent()) {
-                append(STR.".withName(\"\{l.name().get()}\")");
-            }
-        }
-
-        private Constant emitFunctionDescField(FunctionDescriptor desc) {
-            incrAlign();
-            indent();
-            final boolean noArgs = desc.argumentLayouts().isEmpty();
-            NamedConstant descConstant = new NamedConstant(FunctionDescriptor.class);
-            append(STR."static final FunctionDescriptor \{descConstant.constantName} =");
-            if (desc.returnLayout().isPresent()) {
-                append("FunctionDescriptor.of(");
-                emitLayoutString(desc.returnLayout().get());
-                if (!noArgs) {
-                    append(",");
-                }
-            } else {
-                append("FunctionDescriptor.ofVoid(");
-            }
-            if (!noArgs) {
-                append("\n");
-                incrAlign();
-                String delim = "";
-                for (MemoryLayout e : desc.argumentLayouts()) {
-                    append(delim);
-                    indent();
-                    emitLayoutString(e);
-                    delim = ",\n";
-                }
-                append("\n");
-                decrAlign();
-                indent();
-            }
-            append(");\n");
-            decrAlign();
-            return descConstant;
-        }
-
-        private Constant emitConstantString(Object value) {
-            NamedConstant segConstant = new NamedConstant(MemorySegment.class);
-            appendIndentedLines(STR."""
-                static final MemorySegment \{segConstant.constantName} =
-                        RuntimeHelper.CONSTANT_ALLOCATOR.allocateFrom("\{Utils.quote(Objects.toString(value))}");
-                """);
-            return segConstant;
-        }
-
-        private Constant emitConstantAddress(Object value) {
-            NamedConstant segConstant = new NamedConstant(MemorySegment.class);
-            appendIndentedLines(STR."""
-                static final MemorySegment \{segConstant.constantName} =
-                        MemorySegment.ofAddress(\{((Number)value).longValue()}L);
-                """);
-            return segConstant;
-        }
-
-        private Constant emitSegmentField(String nativeName, MemoryLayout layout) {
-            Constant layoutConstant = addLayout(layout);
-            NamedConstant segConstant = new NamedConstant(MemorySegment.class);
-            appendIndentedLines(STR."""
-                static final MemorySegment \{segConstant.constantName} =
-                        RuntimeHelper.lookupGlobalVariable("\{nativeName}", \{layoutConstant});
-                """);
-            return segConstant;
-        }
-
-        String newConstantName() {
-            return "const$" + constantIndex++;
-        }
-    }
-
-    // public API
-
-    public Constant addLayout(MemoryLayout layout) {
-        Constant constant = cache.get(layout);
-        if (constant == null) {
-            constant = builder().emitLayoutField(layout);
-            cache.put(layout, constant);
-        }
-        return constant;
-    }
-
-    public Constant addFieldVarHandle(String nativeName, GroupLayout parentLayout, List<String> prefixElementNames) {
-        return builder().emitFieldVarHandle(nativeName, parentLayout, prefixElementNames);
-    }
-
-    public Constant addGlobalVarHandle(ValueLayout valueLayout) {
-        record VarHandleKey(ValueLayout valueLayout) { }
-        VarHandleKey key = new VarHandleKey(valueLayout.withoutName());
-        Constant constant = cache.get(key);
-        if (constant == null) {
-            constant = builder().emitVarHandle(valueLayout);
-            cache.put(key, constant);
-        }
-        return constant;
-    }
-
-    public Constant addDowncallMethodHandle(String nativeName, FunctionDescriptor descriptor, boolean isVarargs) {
-        return builder().emitDowncallMethodHandleField(nativeName, descriptor, isVarargs, false);
-    }
-
-    public Constant addVirtualDowncallMethodHandle(FunctionDescriptor descriptor) {
-        record DowncallKey(FunctionDescriptor desc) { }
-        DowncallKey downcallKey = new DowncallKey(descriptor);
-        Constant constant = cache.get(downcallKey);
-        if (constant == null) {
-            constant = builder().emitDowncallMethodHandleField(null, descriptor, false, true);
-            cache.put(downcallKey, constant);
-        }
-        return constant;
-    }
-
-    public Constant addUpcallMethodHandle(String className, String name, FunctionDescriptor descriptor) {
-        return builder().emitUpcallMethodHandleField(className, name, descriptor);
-    }
-
-    public Constant addSegment(String nativeName, MemoryLayout layout) {
-        return builder().emitSegmentField(nativeName, layout);
-    }
-
-    public Constant addFunctionDesc(FunctionDescriptor desc) {
-        Constant constant = cache.get(desc);
-        if (constant == null) {
-            constant = builder().emitFunctionDescField(desc);
-            cache.put(desc, constant);
-        }
-        return constant;
-    }
-
-    public Constant addConstantDesc(Class<?> type, Object value) {
-        record ConstantKey(Class<?> type, Object value) { }
-        var key = new ConstantKey(type, value);
-        Constant constant = cache.get(key);
-        if (constant == null) {
-            if (value instanceof String) {
-                constant = builder().emitConstantString(value);
-            } else if (type == MemorySegment.class) {
-                constant = builder().emitConstantAddress(value);
-            } else {
-                constant = ImmediateConstant.ofLiteral(type, value);
-            }
-            cache.put(key, constant);
-        }
-        return constant;
-    }
-}
